Initial Commit
This is the initial commit for the optimization framework. Much of the structure still depends on FiberSim and how we were running things with that model, but this is a start.

Namely, the main driver for the worker class, `worker.fit_worker` relies directly on running FiberSim. This should instead be switched by a generic wrapper function that allows the user to interface with their own model.